chatwave-backend/requirements.txt
django
boto3
django-cors-headers
djangorestframework
requests
PyJWT
cryptography
